[BUG] Page tree disappears when using RecordPicker in TCEforms
The page tree disapperas as soon as you click "Add image" e.g.
in a tt_content element and want to add a media element, so
the vidi based record picker appears.

This is due to the creapy "hide navigation panel code" which currently
is needed as we do not use a common navigation component panel.

Fixes: #31807